Complete auth api specs
bookshelf-touch was updated to the latest version so that the
`updated_at` timestamp of a record is not automatically updated on new
records, which was making tests fail.

`expectErrors` is a new test expectation that checks that a response
from the server is a correctly formatted error response and that it
contains the expected error or errors.