User api specs
The `GET /api/users/:id` route has been updated to perform
authentication before attempting to fetch the user (that way an
unauthorized user will get a 401 Unauthorized response before knowing
whether a user exists or not).

A couple of workarounds have been applied to the Abstract bookshelf
model. By default, bookshelf uses `returning "id"` in its insert
queries, but since we have `uuid` columns with default values generated
in the database, we need `returning "id", "api_id"` or `returning *` so
that we get the `api_id` back when saving models.