Automatically run tests when code changes TG-85
3 new npm scripts:

* `npm run dev:test` runs the tests and then re-runs them every time the
  code changes. This is not automatically launched when running `npm run
  dev`, because reading the test output mixed with the documentation
  generation output is not humanly possible.
* `npm run dev:all` does `npm run dev` + `npm run dev:test`, in case you
  are superhuman and want to do it anyway.
* `npm run test:watch` to automatically run the tests when code changes.

nyc has been configured to save its test coverage report in
`docs/coverage` rather than `coverage`, that way the coverage report is
in the same directory as the rest of the documentation, and is served
along with the API & source code documentation. A link to the coverage
report has been added to the documentation's home page.