Implement POST /api/locations route
A migration was necessary to fix the type of the `locations` table's
`geometry` column, which is now a PostGIS geometry column instead of a
PostgreSQL point.

The new `bookshelf-geojson` dependency handles serializing and
deserializing GeoJSON columns when communicating with PostgreSQL through
Bookshelf.

The new `knex-postgis` dependency was added to avoid issues with
`returning *` in SQL queries returning locations. For the
`bookshelf-geojson` library to be able to parse the location's geometry,
it must be transformed first with `knex-postgis`'s `st.asGeoJSON`
function.

New validators have been added to validate a GeoJSON point (i.e. the
location's `geometry` property).

The workaround that was implemented in `server/models/abstract.js` has
been encapsulated into a Bookshelf plugin and moved to
`server/utils/bookshelf-returning.js`. Ideally, the fix should be made
in Bookshelf itself as a pull request.

The two new Abstract model `parseFrom` and `serializeTo` methods are
shortcuts to parse JSON data into a model's columns, and to serialize a
model's column to JSON. They handle the conversion to and from
underscore and camel-case (using the new `inflection` dependency), and
simplify parsing/serializing deep objects.

Story: TG-43
Tasks: TG-88 TG-89 TG-92