Generate points in bounding box with padding
This is to facilitate end-to-end tests in the frontend. To ensure that
no markers appear out of the bounds of the screen when displaying a map,
the window width/height ratio can be adjusted to the be the same as the
bounding box, and a padding can be applied to ensure that markers that
would be at the very edge of the screen are well within the screen
instead.