Add themes and actions to the database and models
Create two migrations for the `themes` and `actions` tables.
Note that the `actions` table is not fully implemented.
The database diagram has been updated accordingly.

A new section on the `DEVELOPMENT.md` file lists the things that remains to be
done on the currently migrated tables.

There was a bug when logging migrations when more than one were executed
at the same time :
Each migrations was logged one more time than its predecessor.
This commit fixes that bug.

An utility function for creating the touch columns (`created_at` and
`updated_at`) has been created and used in the migrations.
Another function has been created regarding the assertions made in the
test on these two columns. It's used in the tests.

The models for both the tables have been created, along with their
respective tests.

Tasks : TG-112, TG-113, TG-133, TG-134