Integrate orm-query-builder for complex queries (#11)
Integrate orm-query-builder for complex queries

This commit mostly adds the new ORM Query Builder library and uses it
for the bbox filter when retrieving Locations through the API.
Utilities have also been put in place to support pagination, although
this feature is not yet used in this commit. Utility functions to create
query builders have been added to the Abstract model for convenience.

The following libraries have been added to write tests for the new
functionality: chai-as-promised, sinon and sinon-chai.

An unrelated modification: knex has been updated, and the previously
used knex logger function has been removed and replaced by a simpler
version.