Handle method not allowed on routes
Added a new mechanism to handle methods that are not allowed on a specific
route.

This introduces several things:

A new error, called `methodNotAllowed` that sends a 405 HTTP code with
the appropriate header.

This error is called by a middleware, `allowsOnylMethod` in the
`server/api/utils/api.js` file.

This middleware is then used in each `*.routes.js` file, which have
changed a little.
Now, they exports two things :
* The router for this resource
* An object listing the methods allowed on each route

New tests have been added on the test suite for each resource API that
tests that it correctly reject any request on a route with an unallowed
HTTP methods.