Add an API route to list actions (#18)
Add a new `/actions` route to the API. It allows users to list
all available actions of the BioPocket application, without requiring any
authentification. This route implements pagination and inclusion.

Add several validations to the project : one to validate that
a value is an positive integer, one to validate that a value is included
in a whitelist of values, and the last one to validate pagination params
such as offset and limit.
Add tests for the last two validators.

Update the sample-data script to include a dozen themes and hundred
actions related to those themes.

Refactored ApiError implementation. There is now a dedicated file for this
class. Plus it can now contains specific headers that will be added to the
response's headers.
A new badRequest error has been added to the `errors.js` file.

Updated the documentation to include the new API route.
Add new documentation type: `validationError`. It represents validation
errors returned by the routes. It should contain all properties that an
error object can have, regardless of the route.
Also add documentation type for `Action` and `Theme`.

Add two new documentation traits.
`paginatedResource` that documents everything related to pagination in
routes.
`resourceWithInclusion` that documents everything related to resources
that can include other resources in their response objects.

Update previous documentation to include the new types and/or traits.

Add a `.gitattributes` file to force using LF instead of CRLF in all
files.

STORY: TG-52