Create a `data.js` util file.
It should contains utility functions related to data(base) manipulation.
Right now, only the create function is present. It's used in the
sample-data.js script.

This function ought to be used in the mobile e2e tests also,
for creating fake data in the test database.

Also, update the `start:e2e` script so that the backend changes
the limit when retrieving actions ; this is used in the mobile e2e test.