Adds registration management to the API.
Registration being the process of someone registering with their
email to receive updates, notifications or news on the BioPocket project.

Adds a new table `registrations` to store the people information such as
first and last name, and their email address.

Adds a new route `/api/registration` that creates a new registration on
the database.

Add documentation and test for this new API route.

STORY: TG-152